requirements.txt
scapy==2.4.5
tqdm==4.64.1
rich==12.6.0
pysmb==1.2.1